+++ R/aaa.R
@@ -1,0 +1,3 @@
+#' @import rlang
+#' @importFrom purrr map map_int map2 transpose simplify
+NULL