+++ NEWS.md
@@ -1,3 +1,6 @@
+profvis (development version)
+=============================
+
 profvis 0.3.7
 =============